Added the CSSValueList class and made CSSSlashedValue and CSSFunction subtypes thereof. Additionally, CSSColor is now a CSSFunction. Also fixed the problem of CSSDocument->getAllValues() not returning values inside slashed value blocks. CSSSize values inside color function arguments now return false when asked CSSSize->isSize().
These changes may pave the way for a complete overhaul of how the generated structure will look like. Instead of using nested arrays to represent a list of values, each CSSRule could have one value object which is either a CSSPrimitiveValue or a (nested) set of CSSValueList objects. This change would not be backwards-compatible but would result in a much cleaner parsed code structure.