[BUGFIX] Allow at-rules to be parsed in strict mode
The reverts the change to `CSSList` in
https://github.com/MyIntervals/PHP-CSS-Parser/commit/134f4e62fe8ab9f316425f3c0f480b3f4f52d804
and adds a comment that `null` is an expected return value when the end of
the list (or block) is reached.

It also adds a convention, for now, that a test filename beginning with `=`
should mean that it is initially tested with strict parsing enabled.

Fixes #352