[BUGFIX] Invoke PHIVE-installed tools via shell (#484)
On Windows, the `.phive` directory contains batch files that in turn invoke the
phar installed elsewhere on the system.  To be compatible with both Linux and
Windows, the file must be invoked directly via the shell, and without a file
extension specified (when invoked as well as when installed).

Fixes #479.