[TASK] Prepare the 8.6.0 release
There don't seem to be any more enhancements in the immediate pipeline,
so it would be good to get this out now with the four we have,
most or all of which I think have been requested by users.

(I'm not sure I can do the actual release.
I've never made one before, and don't know what's involved/needed.)