[TASK] Switch our code coverage CI from Codacy to Coveralls
Coveralls does not require an API key, provides a nice GitHub Action,
and it can post code coverage comments to PRs.

Also add a code coverage badge to the README.

Also align the GitHub Action workflow with that of our sister project
Emogrifier.

Also add a Composer script for running the tests with coverage.

Fixes #535
Fixes #299