[TASK] Drop the highest/lowest Composer version switch (#698)
As we don't have any direct production dependencies, it does not make
sense to test both with highest and lowest Composer versions on CI.

We accidentally added this highest/lowest switch with the new coverage
job. This commit removes this again, allowing for sharing the Composer
caches with the main CI job.