[TASK] Add TestCase for Value\Color
This covers expected behaviour for CSS level 3
(some of which is also covered by more general tests).

Also included are some commented-out tests for CSS Color Module Level 4,
support for which is yet to be implemented,
as well as for some syntaxes that should be rejected (but currently are not).

Precursor to resolving #755 and supporting CSS Color Module Level 4.