[BUGFIX] Don't render `rgb` colors with `%` values as hex
The only percentage values that could be reliably converted to hex notation are
0%, 20%, 40%, etc.
It's beyond the scope of this library to do that.

Also add additional tests to confirm parsing of percentage values.
Some of these are commented out,
because the input data would result in rendering in an invalid format.
(The "legacy" syntax does not allow a mixture of `percentage`s and `number`s,
so it would be necessary to implement rendering in the "modern" syntax
to resolve those cases, which is beyond the scope of this PR.)