[DOCS] Integrate the `8.x` changelog into the `main` changelog
Now the changelog up to the latest feature release 8.7.0 is
part of the `main` changelog.

This makes it easier to see what has changed after the last
feature release.

Also, this makes it clear in which release some methods were
deprecated, avoid the impression that methods got both
deprecated and removed in the same release.

Also autoformat the changelog.