[TASK] Drop PHP_CodeSniffer
The GPG keys used to sign the PHARs are no longer available,
which breaks installation of PHP_CodeSniffer (and its fixer),
which breaks the build.

In `main`, we already dropped the tool in #468.

Also update PHPStan as the old version was signed with a key
that is no longer available.