[TASK] Drop the PHIVE-installed development tools
It has turned out that our current set of PHIVE-installed
development tools for the V8.x maintenance branch is no longer
installable due to issues with the GPG keys used for signing
the PHARs.

As the maintenance branch has very little activity anyway now,
relying on PHP linting and the unit tests should suffice,
saving us the hassle with PHIVE on that branch.

(We're still keeping some `.gitignore` entries related to the
tools to make switching branches easier.)

Closes #831
Closes #830