[TASK] Drop `getLineNo()` from the `Renderable` interface
The concept of being able to get rendered is logically independent
from having a position in the input stream.

Hence, both concepts should not be intermingled in the
`Renderable` interface.

(We might want to add a dedicated interface for classes that
are linked to specific points in the input stream later, though.)