[FEATURE] Add a utility class to calculate selector specificity (#1049)
Calculating and caching the specificity of a selector is a
different concern than representing a selector, and it deserves
to be in its own class.

This also helps solve the problem of selectors having keep
their specificity cached and in sync with the selector itself.

(We'll have a later change that changes `Selector` to use the
new class.)