[TASK] Add (and use) a `CSSListItem` type
This allows a single type to be used for the contents of a `CSSList`,
instead of a long list of orred types, and helps with static analysis.

Various `assertInstanceOf()` tests are added to the test cases to confirm that
the list items are of the type expected.

Some `implements` and `exetends` lists are now alphabetically sorted.