[TASK] Add (and use) a `CSSListItem` type (#1212)
This allows a single type to be used for the contents of a `CSSList`,
instead of a long list of orred types, and helps with static analysis.

Various `assertInstanceOf()` tests are added to the test cases to confirm that
the list items are of the type expected.

Some `implements` and `exetends` lists are now alphabetically sorted.

Also don't implement interfaces extended by another that is also implemented

PHP<7.4 does not allow this.
Instead, for clarity,
add a DocBlock comment stating which additional interfaces should be
implemented that are not explicitly listed in the `implements` section.

When our minimum PHP version becomes 7.4 or above, we can revisit this.