[TASK] Add `Positionable` interface and implementing trait
This is for CSS items which have a position in the document.

New methods are added:
- `getLineNumber` to replace `getLineNo`;
- `getColumnNumber` to replace `getColNo`.

These return a nullable `int`, instead of using zero to indicate absence.

The old methods are now deprecated,
but defined in the interface and implemented in the trait.

Note that this change only adds the interface and trait.
It does not modify any classes to actually implement or use these.

Part of #1207.