[TASK] Deconflate `getAllValues()` parameters (#1241)
The `$element` parameter was overloaded with a dual purpose.

A second separate parameter has been added for rule filtering,
which is not actually mutually exclusive with CSS subtree selection.

Since `getAllValues()` is part of the public API,
the method now supports being called with the old or new signatures,
with the old signature being deprecated.

Once the deprecation has been included in the 8.x release branch,
the messiness of supporting the previous API can be removed.

Part of #994.  Also relates to #1230.