[TASK] Add interface `RuleContainer` for `RuleSet`
This covers the maniplation of `Rule`s within the container,
and may be implemented by other classes in future (e.g. #1194).

Note that the naming is consistent with the current codebase,
rather than what the CSS entities are now called:
- `Rule` represents what is now called a "declaration";
- `RuleSet` represents what is now called a "declaration block";
- `DeclarationBlock` represents what is now called a "style rule";
- `CSSListItem` (closely) represents what is now generically called a "rule".

Renaming things is part of a longer-term plan touched on in #1189.