[TASK] Update `RuleSet::comparePositionable` to use new methods (#1283)
`getLineNo` and `getColNo` are deprecated.
When the titled method was extracted, use of the above-mentioned methods was
retained to ease backporting and transition to their replacement counterparts:
`getLineNumber` and `getColumnNumber`,
which differ by returning `null` in the case of 'not set'.

This replaces all instances of calls to `getColNo`.

Part of #974